Refactor: moved schemas to model folder
